+++ ddl/5.4/pdw/OMOPCDM_pdw_5.4_ddl.sql
@@ -394,14 +394,14 @@ IF XACT_STATE() = 1 COMMIT; CREATE TABLE  @cdmDatabaseSchema.CONDITION_ERA  (con
 WITH (DISTRIBUTION = HASH(person_id));
 
 --HINT DISTRIBUTE ON KEY (person_id)
-IF XACT_STATE() = 1 COMMIT; CREATE TABLE  @cdmDatabaseSchema.EPISODE  (episode_id bigint NOT NULL,
-			 person_id bigint NOT NULL,
+IF XACT_STATE() = 1 COMMIT; CREATE TABLE  @cdmDatabaseSchema.EPISODE  (episode_id integer NOT NULL,
+			 person_id integer NOT NULL,
 			episode_concept_id integer NOT NULL,
 			episode_start_date date NOT NULL,
 			episode_start_datetime datetime NULL,
 			episode_end_date date NULL,
 			episode_end_datetime datetime NULL,
-			episode_parent_id bigint NULL,
+			episode_parent_id integer NULL,
 			episode_number integer NULL,
 			episode_object_concept_id integer NOT NULL,
 			episode_type_concept_id integer NOT NULL,
@@ -410,8 +410,8 @@ IF XACT_STATE() = 1 COMMIT; CREATE TABLE  @cdmDatabaseSchema.EPISODE  (episode_i
 WITH (DISTRIBUTION = HASH(person_id));
 
 --HINT DISTRIBUTE ON RANDOM
-IF XACT_STATE() = 1 COMMIT; CREATE TABLE  @cdmDatabaseSchema.EPISODE_EVENT  (episode_id bigint NOT NULL,
-			event_id bigint NOT NULL,
+IF XACT_STATE() = 1 COMMIT; CREATE TABLE  @cdmDatabaseSchema.EPISODE_EVENT  (episode_id integer NOT NULL,
+			event_id integer NOT NULL,
 			episode_event_field_concept_id integer NOT NULL )
 WITH (DISTRIBUTION = REPLICATE);
